@@ -17,7 +17,7 @@
         <jar.manifest.mainClass>com.bewsoftware.mdj.cli.Main</jar.manifest.mainClass>
         <docs.source>src/docs/manual</docs.source>
         <docs.target>target/docs/manual</docs.target>
-        <java.version>18</java.version>
+        <java.version>20</java.version>
         <source.encoding>UTF-8</source.encoding>
         <javadoc.source.version>${java.version}</javadoc.source.version>
         <jdk.version>${java.version}</jdk.version>
